data_collector/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
iexfinance==0.5.0
numpy==1.20.1
pandas==1.2.2
praw==7.2.0
prawcore==2.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
update-checker==0.18.0
urllib3==1.26.3
websocket-client==0.57.0
